+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Confirmed Python 3.11 support. ([#39])
 - Added test summary to the end of the test run. ([#31])
 - Added support for `--no-header` and `--no-summary` command line options. ([#64])
+- Added support for capturing terminal output using Rich's `Console` class, with command line option `--rich-capture`. ([#65])
 
 ## [0.1.1] - 2022-03-03
 
@@ -44,3 +45,4 @@ Initial release!
 [#39]: https://github.com/nicoddemus/pytest-rich/pull/39
 [#40]: https://github.com/nicoddemus/pytest-rich/pull/40
 [#64]: https://github.com/nicoddemus/pytest-rich/pull/64
+[#65]: https://github.com/nicoddemus/pytest-rich/pull/65